WoT drawCycle C1: extract draw() body verbatim to table/drawCycle.ts runDrawCycle
Table.draw() collapses to `runDrawCycle(this, fastDraw); return this;`. Free
function reaches the instance via its public surface + get deps() (never #deps),
like cellAccess/domScaffold. Behavior byte-identical: literal if/else + every
isMaster/is() guard preserved; only the receiver is rebound (this -> table,
this.#deps -> table.deps). Dropped two dead totalRows/totalColumns locals.
Gates: test:walkontable 749/0/2, eslint, test:types green.
