+++ handsontable/src/3rdparty/walkontable/src/table.ts
@@ -10,15 +10,13 @@ import {
   isVisible,
 } from '../../../helpers/dom/element';
 import { mixin } from '../../../helpers/object';
-import ColumnFilter from './filter/column';
-import RowFilter from './filter/row';
+import type ColumnFilter from './filter/column';
+import type RowFilter from './filter/row';
 import { Renderer } from './render';
 import ColumnUtils from './utils/column';
 import RowUtils from './utils/row';
-import {
-  CLONE_BOTTOM,
-  CLONE_BOTTOM_INLINE_START_CORNER,
-} from './overlay';
+import { CLONE_BOTTOM } from './overlay';
+import { runDrawCycle } from './table/drawCycle';
 
 /**
  * Assembles the Table module's dependencies from the engine composition context. Shared by the
@@ -340,180 +338,7 @@ class Table {
    * @returns {Table}
    */
   draw(fastDraw = false) {
-    const { wtSettings } = this;
-    const wtOverlays = this.#deps.getWtOverlays();
-    const wtViewport = this.#deps.getWtViewport();
-    const totalRows = wtSettings.getSetting('totalRows');
-    const totalColumns = wtSettings.getSetting('totalColumns');
-    const rowHeaders = wtSettings.getSetting<Function[]>('rowHeaders');
-    const rowHeadersCount = rowHeaders.length;
-    const columnHeaders = wtSettings.getSetting<Function[]>('columnHeaders');
-    const columnHeadersCount = columnHeaders.length;
-    let runFastDraw = fastDraw;
-
-    if (this.isMaster) {
-      wtOverlays.beforeDraw();
-      this.holderOffset = this.#deps.geometryReader.offset(this.holder);
-
-      wtViewport.rowHeightCache.ensureBuilt();
-      wtViewport.columnWidthCache.ensureBuilt();
-
-      // S16b: on the single-pass gated path, decide and apply the header-border classes
-      // (`innerBorderTop` / `innerBorderInlineStart`) BEFORE resolving the snapshot and rendering the
-      // cells, from the pre-render scroll position + settings. Cells then render in their final
-      // position, so the post-render `resetFixedPosition` toggle is a no-op (`positionChanged` stays
-      // `false`) and the nested `wot.draw(true)` re-render never fires. Element mode only (guaranteed
-      // by the gate); the border box is thus present when `beginDrawLayout` measures the workspace,
-      // matching a steady-state scrolled draw.
-      if (wtViewport.usesLayoutSnapshotForCalculators()) {
-        wtOverlays.prepareHeaderBorders();
-      }
-
-      // Resolve the single-pass layout snapshot for this draw (scrollbar prediction from numbers).
-      // Not yet the source of truth for the calculators below — see Viewport#beginDrawLayout.
-      wtViewport.beginDrawLayout();
-
-      runFastDraw = wtViewport.createCalculators(runFastDraw);
-
-      if (rowHeadersCount && !wtSettings.getSetting('fixedColumnsStart')) {
-        const leftScrollPos = wtOverlays.inlineStartOverlay.getScrollPosition();
-        const previousState = this.correctHeaderWidth;
-
-        this.correctHeaderWidth = leftScrollPos !== 0;
-
-        if (previousState !== this.correctHeaderWidth) {
-          runFastDraw = false;
-        }
-      }
-    }
-
-    if (runFastDraw) {
-      if (this.isMaster) {
-        wtOverlays.refresh(true);
-        // Fast (scroll) draws skip the full-render header-height pass below, so the master/top
-        // header heights can drift against the frozen overlays during scrolling (a tall wrapped
-        // frozen header that the master never renders). Re-sync here. The method is a cheap no-op
-        // unless the grid has frozen columns with column headers, so non-frozen grids are unaffected.
-        this.syncOversizedColumnHeadersWithFrozenOverlays();
-      }
-    } else {
-      if (this.isMaster) {
-        this.tableOffset = this.#deps.geometryReader.offset(this.TABLE);
-      } else {
-        this.tableOffset = this.#deps.getParentTableOffset();
-      }
-
-      const startRow = Math.max(this.getFirstRenderedRow(), 0);
-      const startColumn = Math.max(this.getFirstRenderedColumn(), 0);
-
-      this.rowFilter = new RowFilter(startRow, wtSettings.getSetting<number>('totalRows'), columnHeadersCount);
-      this.columnFilter = new ColumnFilter(startColumn, wtSettings.getSetting<number>('totalColumns'), rowHeadersCount);
-
-      let performRedraw = true;
-
-      // Only master table rendering can be skipped
-      if (this.isMaster) {
-        this.alignOverlaysWithTrimmingContainer(); // todo It calls method from child class (MasterTable).
-        const skipRender: { skipRender?: boolean } = {};
-
-        this.wtSettings.getSetting('beforeDraw', true, skipRender);
-        performRedraw = skipRender.skipRender !== true;
-      }
-
-      if (performRedraw) {
-        this.tableRenderer.setHeaderContentRenderers(rowHeaders, columnHeaders);
-
-        if (this.is(CLONE_BOTTOM) ||
-            this.is(CLONE_BOTTOM_INLINE_START_CORNER)) {
-          // do NOT render headers on the bottom or bottom-left corner overlay
-          this.tableRenderer.setHeaderContentRenderers(rowHeaders, []);
-        }
-
-        this.resetOversizedRows();
-
-        this.tableRenderer
-          .setActiveOverlayName(this.name)
-          .setViewportSize(this.getRenderedRowsCount(), this.getRenderedColumnsCount())
-          .setFilters(this.rowFilter, this.columnFilter)
-          .render();
-
-        this.adjustColumnHeaderHeights();
-
-        if (this.isMaster || this.is(CLONE_BOTTOM)) {
-          this.markOversizedRows();
-        }
-
-        if (this.isMaster) {
-          if (!this.wtSettings.getSetting('externalRowCalculator')) {
-            // Single-pass: the fully/partially-visible calculators were already computed in pass 1
-            // (`createCalculators`) from the layout snapshot. On the snapshot path, the only thing that
-            // can change them post-render is `markOversizedRows` invalidating the row-height cache when
-            // rendered content is genuinely taller than its configured size. If both size caches are
-            // still current, the post-render values are identical, so this second pass is a redundant
-            // recompute — skip it. (Read `isCurrent()` before `ensureBuilt()` rebuilds an invalidated
-            // cache. Validated non-destructively across the full e2e: this predicate would skip 595
-            // draws with byte-identical ranges and 0 divergences.) The legacy measured path, and any
-            // draw with an oversized row, still recompute here.
-            const skipSecondPass = wtViewport.usesLayoutSnapshotForCalculators() &&
-              wtViewport.rowHeightCache.isCurrent() && wtViewport.columnWidthCache.isCurrent();
-
-            wtViewport.rowHeightCache.ensureBuilt();
-            wtViewport.columnWidthCache.ensureBuilt();
-
-            if (!skipSecondPass) {
-              wtViewport.createVisibleCalculators();
-            }
-          }
-
-          wtOverlays.refresh(false);
-          this.syncOversizedColumnHeadersWithFrozenOverlays();
-          wtOverlays.applyToDOM();
-
-          this.wtSettings.getSetting('onDraw', true);
-
-        } else if (this.is(CLONE_BOTTOM)) {
-          this.#deps.getCloneSource().wtOverlays.adjustElementsSize();
-        }
-      }
-    }
-
-    let positionChanged = false;
-
-    if (this.isMaster) {
-      positionChanged = wtOverlays.topOverlay.resetFixedPosition();
-
-      if (wtOverlays.bottomOverlay.clone) {
-        positionChanged = wtOverlays.bottomOverlay.resetFixedPosition() || positionChanged;
-      }
-
-      positionChanged = wtOverlays.inlineStartOverlay.resetFixedPosition() || positionChanged;
-
-      if (wtOverlays.topInlineStartCornerOverlay) {
-        wtOverlays.topInlineStartCornerOverlay.resetFixedPosition();
-      }
-
-      if (wtOverlays.bottomInlineStartCornerOverlay && wtOverlays.bottomInlineStartCornerOverlay.clone) {
-        wtOverlays.bottomInlineStartCornerOverlay.resetFixedPosition();
-      }
-    }
-
-    if (positionChanged) {
-      // It refreshes the cells borders caused by a 1px shift (introduced by overlays which add or
-      // remove `innerBorderTop` and `innerBorderInlineStart` CSS classes to the DOM element. This happens
-      // when there is a switch between rendering from 0 to N rows/columns and vice versa).
-      wtOverlays.refreshAll(); // `refreshAll()` internally already calls `refreshSelections()` method
-      wtOverlays.adjustElementsSize();
-    } else {
-      this.#deps.getSelectionManager()
-        .setActiveOverlay(this.facadeGetter())
-        .render(runFastDraw);
-    }
-
-    if (this.isMaster) {
-      wtOverlays.afterDraw();
-    }
-
-    this.#deps.setDrawn(true);
+    runDrawCycle(this, fastDraw);
 
     return this;
   }
